requirements.txt
sqlalchemy
sqlalchemy_utils
requests
watchdog
psycopg2-binary
pyyaml
flask
flask-cors
flask-sqlalchemy
flask-restplus
flask-jwt-extended
bcrypt
python-dateutil
unidecode
uwsgi